cleanup ai slop
